update testsuite to use defmt-test
- Add tests for Cortex-M0 as well
