+++ testsuite/README.md
@@ -1,68 +1,54 @@
 # Testsuite
 
-This workspace contains tests that run on physical and simulated Cortex-M CPUs.
-
-## Building
-
-Exactly one of these features are required:
-
-* `semihosting` Use semihosting for logging, this is used for QEMU.
-* `rtt` Use RTT for logging, this is used with physical cortex-m CPUs.
-
-Assuming you are at the root of the repository you can build like this:
-
-```console
-$ cd testsuite
-$ cargo build
-   Compiling testsuite v0.1.0 (cortex-m/testsuite)
-    Finished dev [unoptimized + debuginfo] target(s) in 0.08
-```
+This workspace contains tests that run on physical and simulated Cortex-M CPUs. It uses
+the [`defmt-test`](https://github.com/knurling-rs/defmt/tree/main/firmware/defmt-test) library to
+do this.
 
 ## Running with QEMU
 
 The runner is already configured for QEMU in `testsuite/.cargo/config.toml`.
-Use the `semihosting` feature for logging, QEMU does not have native support for RTT.
+You need to install `qemu-run` and `qemu-system-arm`.
+
+For example, on Ubuntu, you can use:
+
+```sh
+sudo apt install qmu-system-arm
+cargo install qemu-run
+```
+
+`qemu-run` is a wrapper around `qemu-system-arm` which also processes the `defmt` logs properly.
+You also need to activate the `qemu` feature when running the tests.
 
 For more information on QEMU reference the QEMU section in [The Embedded Rust Book].
 
+*Cortex-M3*
+
 ```console
 $ cd testsuite
-$ cargo run
-    Finished dev [unoptimized + debuginfo] target(s) in 0.01s
-     Running `qemu-system-arm -cpu cortex-m3 -machine lm3s6965evb -nographic -semihosting-config enable=on,target=native -kernel /cortex-m/target/thumbv7m-none-eabi/debug/testsuite`
-Timer with period zero, disabling
-Hello world!
-(1/1) running `double_take`...
-all tests passed!
+$ cargo test --features qemu --target thumbv7em-none-eabihf
+```
+
+*Cortex-M0*
+
+```console
+$ cd testsuite
+$ cargo test --features qemu --target thumbv6m-none-eabi --release
 ```
 
 ## Running with Physical Hardware
 
-No implementation-specific features are tested right now; any physical `thumbv7m` target should work.
-
 Tests are executed with [probe-rs](https://github.com/probe-rs/probe-rs).
 
-* Update `memory.x` in the root of the repository to match your target memory layout.
-* Change the `probe-rs` chip argument to match your chip, supported chips can be found with `probe-rs chip list`
+* Create or update `memory.x` in the `testsuite` directory to match your target memory layout.
+* Change the `.cargo/config.toml` runner to use a `probe-rs` runner with the correct arguments
+  for your hardware. You can find support chips with `probe-rs chip list`.
+* Alternatively, you can set a target specific runner using the `CARGO_TARGET_<TARGET_TRIPLE>_RUNNER`
+  environmental variable.
 * Change the target to match your CPU
 
 ```console
-$ sed -i 's/FLASH : ORIGIN = 0x00000000, LENGTH = 256K/FLASH : ORIGIN = 0x8000000, LENGTH = 256K/g' memory.x
 $ cd testsuite
-$ cargo build --target thumbv7em-none-eabi --features rtt
-   Compiling minitest v0.1.0 (/cortex-m/testsuite/minitest)
-   Compiling testsuite v0.1.0 (/cortex-m/testsuite)
-    Finished dev [unoptimized + debuginfo] target(s) in 0.16s
-$ probe-rs run --chip STM32WLE5JCIx --connect-under-reset ../target/thumbv7em-none-eabi/debug/testsuite
-(HOST) INFO  flashing program (19 pages / 19.00 KiB)
-(HOST) INFO  success!
-────────────────────────────────────────────────────────────────────────────────
-Hello world!
-(1/2) running `double_take`...
-(2/2) running `cycle_count`...
-all tests passed!
-────────────────────────────────────────────────────────────────────────────────
-(HOST) INFO  device halted without error
+$ cargo test --features hardware
 ```
 
 ## License
