--- testsuite/minitest/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "minitest"
-publish = false
-edition = "2018"
-version = "0.1.0"
-
-[features]
-rtt = ["rtt-target", "minitest-macros/rtt"]
-
-[dependencies]
-cortex-m-rt.path = "../../cortex-m-rt"
-cortex-m-semihosting.path = "../../cortex-m-semihosting"
-cortex-m.path = "../../cortex-m"
-minitest-macros.path = "macros"
-rtt-target = { version = "0.5.0", optional = true }
