--- testsuite/minitest/README.md
@@ -1,7 +1,0 @@
-# mini-test
-
-This is an embedded test framework forked from knurling's excellent [`defmt-test`] crate.
-
-This even more minimal than [`defmt-test`] to allow for for testing of this crate without dependency cycles.
-
-[`defmt-test`]: https://crates.io/crates/defmt-test/
