--- testsuite/minitest/macros/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "minitest-macros"
-publish = false
-edition = "2018"
-version = "0.1.0"
-
-[lib]
-proc-macro = true
-
-[features]
-rtt = []
-
-[dependencies]
-proc-macro2 = "1.0.29"
-quote = "1.0.10"
-syn = { version = "2.0.68", features = ["extra-traits", "full"] }
